flask_app/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
Flask-MySQL==1.6.0
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector==2.2.9
mysql-connector-python==9.2.0
packaging==24.2
PyMySQL==1.1.1
typing_extensions==4.12.2
Werkzeug==3.1.3
